+++ chaincode/src/main/java/de/upb/cs/uc4/chaincode/ContractBase.java
@@ -1,0 +1,21 @@
+package de.upb.cs.uc4.chaincode;
+
+import org.hyperledger.fabric.contract.Context;
+import org.hyperledger.fabric.contract.ContractInterface;
+import org.hyperledger.fabric.contract.annotation.Transaction;
+
+
+public abstract class ContractBase implements ContractInterface {
+
+    private final String version = "v0.10.1";
+
+    /**
+     * Gets version of the chaincode
+     * @param ctx transaction context providing access to ChaincodeStub etc.
+     * @return certificate on success, serialized error on failure
+     */
+    @Transaction()
+    public String getVersion(final Context ctx) {
+        return version;
+    }
+}